// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      courses: {
        Row: {
          code: string
          created_at: string | null
          department: Database["public"]["Enums"]["department"]
          description: string | null
          duration: string
          eligibility: string | null
          id: number
          title: string
          type: Database["public"]["Enums"]["course_type"]
          updated_at: string | null
        }
        Insert: {
          code: string
          created_at?: string | null
          department: Database["public"]["Enums"]["department"]
          description?: string | null
          duration: string
          eligibility?: string | null
          id?: number
          title: string
          type: Database["public"]["Enums"]["course_type"]
          updated_at?: string | null
        }
        Update: {
          code?: string
          created_at?: string | null
          department?: Database["public"]["Enums"]["department"]
          description?: string | null
          duration?: string
          eligibility?: string | null
          id?: number
          title?: string
          type?: Database["public"]["Enums"]["course_type"]
          updated_at?: string | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      check_admin_credentials: {
        Args: { p_email: string; p_password: string }
        Returns: Json
      }
      get_current_user_role: {
        Args: Record<PropertyKey, never>
        Returns: string
      }
    }
    Enums: {
      course_type: "undergraduate" | "postgraduate" | "diploma" | "certificate"
      department:
        | "Science"
        | "Arts & Humanities"
        | "Commerce"
        | "Computer Science"
      event_category: "cultural" | "academic" | "sports" | "career" | "alumni"
      user_role: "admin" | "faculty" | "student"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never